refactor: move record data classes to pika_zoo.records package
Separate data containers (FrameRecord, RoundRecord, GameRecord,
GamesRecord) from RecordGame wrapper into pika_zoo.records.
Add records/README.md with full column reference, player states,
CSV format spec. Slim wrappers/README.md RecordGame section to
link to records docs. Backward-compatible re-exports preserved.
