+++ src/pika_zoo/wrappers/__init__.py
@@ -1,6 +1,7 @@
+from pika_zoo.records import FrameRecord, FrameSnapshot, GameRecord, GamesRecord, RoundRecord
 from pika_zoo.wrappers.convert_single_agent import ConvertSingleAgent
 from pika_zoo.wrappers.normalize_observation import NormalizeObservation
-from pika_zoo.wrappers.record_game import FrameRecord, FrameSnapshot, GameRecord, GamesRecord, RecordGame, RoundRecord
+from pika_zoo.wrappers.record_game import RecordGame
 from pika_zoo.wrappers.reward_shaping import RewardShaping
 from pika_zoo.wrappers.simplify_action import SimplifyAction
 from pika_zoo.wrappers.simplify_observation import SimplifyObservation
